fix: emojis appear differently on different platforms
Include the Noto Color Emoji font in the package and use it to render
emojis. This resolves the missing emojis on Android, and ensures
they will render properly on Linux systems without the font or an
old version.

Windows does not appear to support Noto Color Emoji, so the emojis
will still appear differently on Windows.
See: https://github.com/googlefonts/noto-emoji/issues/43

In addition, the native Windows font only supports Unicode 14.0
emojis, so newer emojis are not shown on Windows.