feat: ctrl + f opens search, esc closes search
Automatically redirecting keys to search when the user began typing
was proving to cause issues with triggering the search unintentionally,
missing key entry if typing too fast, etc.

This change makes it so that the user must press ctrl + f to open the
search, and esc to close it. This should make the search more predictable
and less intrusive.